Fix:금액/비율 타입 BigDecimal로 통일

+++ core-api/src/main/java/com/pocketstock/core/asset/dto/SpendingResponse.java
@@ -1,5 +1,6 @@
 package com.pocketstock.core.asset.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record SpendingResponse(long totalSpending, List<CategorySpending> categories) {}
+public record SpendingResponse(BigDecimal totalSpending, List<CategorySpending> categories) {}
